// src/api/generated/model/getApiV1CatCareCatsCatIdBowelMovements200Item.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */
import type { GetApiV1CatCareCatsCatIdBowelMovements200ItemStoolType } from './getApiV1CatCareCatsCatIdBowelMovements200ItemStoolType';

export type GetApiV1CatCareCatsCatIdBowelMovements200Item = {
  id: string;
  catId: string;
  recordedBy: string;
  recordedAt: string;
  /** @nullable */
  stoolType?: GetApiV1CatCareCatsCatIdBowelMovements200ItemStoolType;
  isAbnormal: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
};